Optionally grant experience from farming and foraging
